chore(nlu): fix linter errors
