+++ packages/botonic-nlu/src/simple-nn.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/unbound-method */
 import {
   Sequential,
   sequential,
