+++ packages/botonic-nlu/tests/nlu/data-reader.test.ts
@@ -11,7 +11,7 @@ describe('Testing Botonic NLU main methods', () => {
     'en',
     'data.csv',
   );
-  it('it loads from a csv correctly', () => {
+  it('loads from a csv correctly', () => {
     const dataReader = new DataReader();
     const data = dataReader.readData(dataPath);
     expect(data).toHaveLength(136);
